Moved provider connection tests in test base class
