+++ tests/src/python/CMakeLists.txt
@@ -11,7 +11,8 @@ ENDIF (WITH_SERVER)
 ADD_PYTHON_TEST(PyCoreAdittions test_core_additions.py)
 ADD_PYTHON_TEST(PyPythonRepr test_python_repr.py)
 ADD_PYTHON_TEST(PyQgsActionManager test_qgsactionmanager.py)
-ADD_PYTHON_TEST(PyQgsProviderConnection test_qgsproviderconnection.py)
+ADD_PYTHON_TEST(PyQgsProviderConnectionPostgres test_qgsproviderconnection_postgres.py)
+ADD_PYTHON_TEST(PyQgsProviderConnectionGpkg test_qgsproviderconnection_ogr_gpkg.py)
 ADD_PYTHON_TEST(PyQgsAFSProvider test_provider_afs.py)
 ADD_PYTHON_TEST(PyQgsAlignmentComboBox test_qgsalignmentcombobox.py)
 ADD_PYTHON_TEST(PyQgsPythonProvider test_provider_python.py)
